add cues state class
