+++ feature/player/src/main/java/dev/anilbeesetti/nextplayer/feature/player/ui/SubtitleView.kt
@@ -1,26 +1,19 @@
 package dev.anilbeesetti.nextplayer.feature.player.ui
 
-import android.content.Context
 import android.graphics.Typeface
 import android.util.TypedValue
 import android.view.accessibility.CaptioningManager
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat.getSystemService
 import androidx.media3.common.Player
-import androidx.media3.common.listen
-import androidx.media3.common.text.Cue
 import androidx.media3.ui.CaptionStyleCompat
 import androidx.media3.ui.SubtitleView
 import dev.anilbeesetti.nextplayer.core.model.PlayerPreferences
 import dev.anilbeesetti.nextplayer.feature.player.extensions.toTypeface
+import dev.anilbeesetti.nextplayer.feature.player.state.rememberCuesState
 
 @Composable
 fun SubtitleView(
@@ -28,14 +21,7 @@ fun SubtitleView(
     player: Player,
     playerPreferences: PlayerPreferences,
 ) {
-    var currentCues by remember { mutableStateOf<List<Cue>>(emptyList()) }
-    LaunchedEffect(Unit) {
-        player.listen { events ->
-            if (events.contains(Player.EVENT_CUES)) {
-                currentCues = player.currentCues.cues
-            }
-        }
-    }
+    val cuesState = rememberCuesState(player)
 
     AndroidView(
         modifier = modifier.fillMaxSize(),
@@ -64,7 +50,7 @@ fun SubtitleView(
             }
         },
         update = { subtitleView ->
-            subtitleView.setCues(currentCues)
+            subtitleView.setCues(cuesState.cues)
         },
     )
 }
